refactor(example): split display into separate thread
Rather than have everything in the same loop and thread, this splits off
a separate thread for drawing to the display, with two atomic reference.
First to a trigger boolean to know when to draw the white square, and
second to a drawing delay to be able to report back the value to the
main thread.