docs(README): fix error in circuit
The previous circuit diagram would not work if you didn't have an
oscilloscope connected to Vout. This is because we leave DC current no
path to ground. This fixes the circuit to match the intention: get a
reading from the photodiode in photoconductive mode while also doing a
little low-pass filtering.